requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pyserial==3.5
jinja2==3.1.4
